requirements.txt
discord.py
yt-dlp
python-dotenv
pynacl
ffmpeg
aiohttp
cloudconvert